Move uop_t to uops.h
